// src/shared/api/generated/models/caseIntelligenceAuthoringIssueDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CaseIntelligenceAuthoringIssueDtoCode } from "./caseIntelligenceAuthoringIssueDtoCode";
import type { CaseIntelligenceAuthoringIssueDtoSeverity } from "./caseIntelligenceAuthoringIssueDtoSeverity";

export interface CaseIntelligenceAuthoringIssueDto {
  code: CaseIntelligenceAuthoringIssueDtoCode;
  /** @maxLength 500 */
  message: string;
  /** @maxLength 256 */
  path: string;
  /** @maxItems 10 */
  relatedPaths: string[];
  severity: CaseIntelligenceAuthoringIssueDtoSeverity;
}